+++ learn_web/webapp/__init__.py
@@ -1,11 +1,11 @@
-from flask import Flask, render_template, flash, url_for, redirect
-from flask_login import LoginManager, current_user, login_required, \
-    login_user, logout_user
+from flask import Flask, render_template
+from flask_login import LoginManager, current_user, login_required
 
-from .forms import LoginForm
 from .weather import weather_by_city
 from .python_org_news import get_python_news
-from .model import db, News, User
+from .model import db, News
+from .user.models import User
+from .user.views import blueprint as user_blueprint
 
 
 def create_app():
@@ -16,7 +16,8 @@ def create_app():
 
     login_manager = LoginManager()
     login_manager.init_app(app)
-    login_manager.login_view = 'login'
+    login_manager.login_view = 'user.login'
+    app.register_blueprint(user_blueprint)
 
     @login_manager.user_loader
     def load_user(user_id):
@@ -29,32 +30,6 @@ def create_app():
         news_list = News.query.order_by(News.published.desc()).all()
         return render_template('index.html', page_title=page_title,
                                weather=weather, news_list=news_list)
-
-    @app.route('/login')
-    def login():
-        if current_user.is_authenticated:
-            return redirect(url_for('index'))
-        title = "Авторизация"
-        login_form = LoginForm()
-        return render_template("login.html", page_title=title, form=login_form)
-
-    @app.route('/process-login', methods=['POST'])
-    def process_login():
-        form = LoginForm()
-        if form.validate_on_submit():
-            user = User.query.filter_by(username=form.username.data).first()
-            if user and user.check_password(form.password.data):
-                login_user(user)
-                flash('Вы успешно вошли на сайт')
-                return redirect(url_for('index'))
-        flash('Неправильное имя или пароль')
-        return redirect(url_for('login'))
-
-    @app.route('/logout')
-    def logout():
-        logout_user()
-        flash('Пока!')
-        return redirect(url_for('index'))
 
     @app.route('/admin')
     @login_required
