+++ learn_web/webapp/templates/login.html
@@ -22,7 +22,7 @@
                         </div>
                     {% endif %}
                 {% endwith %}
-                <form action="{{ url_for('process_login') }}" method="POST">
+                <form action="{{ url_for('user.process_login') }}" method="POST">
                     {{ form.hidden_tag() }}
                     <div class="form-group">
                         {{ form.username.label }}
@@ -40,7 +40,6 @@
                 </form>
             </div>
             <div class="col-8">
-
             </div>
         </div>
     <!-- Optional JavaScript -->
